test: first test setup finished

+++ src/use-cases/sheets/delete.ts
@@ -5,6 +5,12 @@ export class DeleteSheetUseCase {
 	constructor(private sheetsRepository: ISheetsRepository) {}
 
 	async execute(data: IDeleteSheetParams) {
+		const sheet = await this.sheetsRepository.get(data)
+
+		if (!sheet.data) {
+			throw new Error('Sheet not found')
+		}
+
 		await this.sheetsRepository.delete(data)
 	}
 }
